requirements.txt
requests==2.18.4
elasticsearch>=6.0.0,<7.0.0
google-cloud-vision
